backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
channels==4.0.0
channels-redis==4.1.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
kombu==5.6.1
msgpack==1.1.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
